前台模块还差文章详情，后台模块还差统计 Try tomorrow ！

+++ weblog-server/src/main/java/com/disscheng/weblog/service/impl/CategoryServiceImpl.java
@@ -1,24 +1,35 @@
 package com.disscheng.weblog.service.impl;
 
+import com.disscheng.weblog.constant.MessageConstant;
+import com.disscheng.weblog.dto.ArticlePageQueryByCategoryDTO;
+import com.disscheng.weblog.dto.ArticlePageQueryDTO;
 import com.disscheng.weblog.dto.CategoryDTO;
 import com.disscheng.weblog.dto.CategoryPageQueryDTO;
+import com.disscheng.weblog.entity.Article;
 import com.disscheng.weblog.entity.Category;
+import com.disscheng.weblog.exception.ArticleDeleteException;
+import com.disscheng.weblog.mapper.ArticleMapper;
 import com.disscheng.weblog.mapper.CategoryMapper;
 import com.disscheng.weblog.service.CategoryService;
+import com.disscheng.weblog.vo.ArticlePageQueryVO;
 import com.disscheng.weblog.vo.CategoryPageQueryVO;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@Slf4j
 public class CategoryServiceImpl implements CategoryService {
     @Autowired
     private CategoryMapper categoryMapper;
-
+    @Autowired
+    private ArticleMapper articleMapper;
     /**
      * 分页查询分类列表
      * @param categoryPageQueryDTO
@@ -61,6 +72,35 @@ public class CategoryServiceImpl implements CategoryService {
      * @param id
      */
     public void delete(Long id) {
+        if(categoryMapper.getArticles(id) > 0){
+            throw new ArticleDeleteException(MessageConstant.ARTICLE_DELETE_CATEGORY_EXIST);
+        }
         categoryMapper.delete(id);
+    }
+
+    /**
+     * 根据分类查询文章列表
+     * @param articlePageQueryByCategoryDTO
+     * @return
+     */
+    public ArticlePageQueryVO getArticlesByCategory(ArticlePageQueryByCategoryDTO articlePageQueryByCategoryDTO) {
+        PageHelper.startPage(articlePageQueryByCategoryDTO.getCurrent(), articlePageQueryByCategoryDTO.getSize());
+        Page<Long> articleIdList = categoryMapper.getArticleIdListByCategoryId(articlePageQueryByCategoryDTO.getId());
+        List<Article> articleList = new ArrayList<>();
+        for (Long articleId : articleIdList) {
+            Article article = articleMapper.getArticle(articleId);
+            if (article != null) {
+                articleList.add(article);
+            }
+        }
+        log.info("articleList:{}", articleList);
+        ArticlePageQueryVO articlePageQueryVO = ArticlePageQueryVO.builder()
+                .current(articlePageQueryByCategoryDTO.getCurrent())
+               .size(articleIdList.getPageSize())
+               .total(articleIdList.getTotal())
+               .pages(articleIdList.getPages())
+                .data(articleList)
+                .build();
+        return articlePageQueryVO;
     }
 }
